# server/src/swaggerDefinition.json
{
  "swagger": "2.0",
  "info": {
    "title": "Reddit Clone API DOCS",
    "description": "Reddit 클론코딩 API 정리 문서입니다.",
    "version": "1.0.0",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "dlehddnr0713@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:4000",
...
  "schemes": ["http"],
  "paths": {
...
    "/api/subs": {
      "post": {
        "tags": ["subs"],
        "summary": "코멘트 조회 API",
        "operationId": "getComments",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/api/subs/{name}": {
      "get": {
        "tags": ["subs"],
        "summary": "특정 코멘트 조회 API",
        "operationId": "getComments",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  }
}